Optimizations to CSS and JS
Added fallbacks for CSS to all clamp styles for browsers that don't support it. For Javascript I added code that would get the thumbnail image alt description text and supply it to the lightbox image as well, I changed a long "if else else if statement" for the source elements for the lightbox picture to be a more optimized switch statement. Also added "Use Strict" to the top of the code as I forgot to do that before and it was causing errors in Safari where it didn't recognize some variables in the gallery code.