Improvements to CSS and Javascript
Changed the html font size in CSS to 62.5% from 10px to allow for scalability.

Edited the Javascript CloseNav function to remove the event listener that listens for keyboard input of "escape" key.

moved the left and right lightbox button event listeners as well as the event listener for whether a user has clicked outside of the lightbox and to close it if so, to inside of the closeLightbox function.

Basic improvements to remove unnecessary event listeners when they aren't needed in the moment.